rename component.jsx files
