+++ src/web/pages/groups/GroupComponent.jsx
